+++ lib/index.js
@@ -14,6 +14,8 @@ module.exports.ast = require('./ast');
 module.exports.parse = Parser.parse.bind(Parser);
 module.exports.SyntaxError = Parser.SyntaxError;
 
+module.exports.contains_func = require('./astutil').contains_func;
+
 var check = module.exports.check = function(input, options) {
     /* status is one character:
      *  + : success! result is in 'output'
